Break out ListBox.ItemArray nested classes
